py-srv/bin/requirements.txt
bottle
bottle-sqlalchemy
psycopg2

# for ssl
cheroot

# for logins
beaker
requests